finish hook and fix a bug in fd_manager::get

+++ sylar/hook.h
@@ -81,7 +81,7 @@ extern "C" {
   typedef int (*fcntl_fun) (int fd, int cmd, ...);
   extern fcntl_fun fcntl_f;
 
-  typedef int (*ioctl_fun) (int d, int request, ...);
+  typedef int (*ioctl_fun) (int d, unsigned long request, ...);
   extern ioctl_fun ioctl_f;
 
   typedef int (*getsockopt_fun) (int socket, int level, int option_name,
